perf(signals): keep A28 promotion off the recompute hot path
CodSpeed flagged 13 benchmarks on #3337 (up to -13.9%). Root cause: every
recompute called unflushedCursor() + promoteUnflushed(from) through the
scheduler module, and promoteUnflushed truncated an empty list on each
run (unflushedNodes.length = from is a runtime call, not inlined).

- Move the unflushed list, markUnflushed and promoteUnflushed into
  core.ts so recompute compares the length locally and only calls
  promote when the run actually wrote something.
- promoteUnflushed returns before truncating an empty list.
- Plain (non-optimistic) nodes skip the hasActiveOverride probe.

Measured (dev tier, update1to1): base 0.65 ms, #3337 head 0.87, fixed
0.71. diamond and update1to1000 back to parity; the residual on the
one-write-one-memo path is the deferred subscriber walk A28 requires
(one extra pass over written nodes per flush) and is the price of
"writes visible at flush", not an implementation cost.

Size: raw floor +27 B (22,252 -> 22,279); brotli scenarios move by up
to +130 B from compression reordering of the moved block.
