+++ packages/signals/src/core/async.ts
@@ -15,7 +15,7 @@ import {
   STATUS_UNINITIALIZED
 } from "./constants.js";
 import { attrHooks } from "./attribution-hooks.js";
-import { context, setSignal, untrack, ext, statusNotifierOf } from "./core.js";
+import { context, setSignal, untrack, ext, statusNotifierOf, markUnflushed } from "./core.js";
 import { devTrackHeldPending } from "./invariants.js";
 import { emitDiagnostic } from "./dev.js";
 import { NotReadyError, StatusError } from "./error.js";
@@ -33,7 +33,6 @@ import {
   globalQueue,
   insertSubs,
   origin,
-  markUnflushed,
   queuePendingNode,
   schedule,
   setOrigin,
